turns out that we need to equip before choosing powers after all.
This makes it a little harder to acquire new weapon/armor proficiencies,
since the character can wind up with "orphaned" powers (which no longer
are useful with the new weapon). But there are some powers that depend on
weapons of particular types being wielded, and those can't be selected
until a weapon is wielded. So, catch 22!